+++ src/TwilioVideo.android.js
@@ -57,7 +57,7 @@ const propTypes = {
   /**
    * Called when requesting a room snapshot via fetchRoom()
    *
-   * @param {{sid?: string, name?: string, mediaRegion?: string, state?: string, localParticipant: Participant, remoteParticipants: Participant[], dominantSpeaker?: Participant}}
+   * @param {{sid?: string, name?: string, mediaRegion?: string, state?: string, localParticipant: Participant, signalingRegion?: string, remoteParticipants: Participant[], dominantSpeaker?: Participant}}
    */
   onRoomFetched: PropTypes.func,
   /**
@@ -370,6 +370,7 @@ class CustomTwilioVideoView extends Component {
    * @param {Object} params - Connection parameters
    * @param {string} params.roomName - The room name to connect to
    * @param {string} params.accessToken - The Twilio JWT access token
+   * @param {string} params.region - The Twilio Signaling Region to connect to
    * @param {'front'|'back'} [params.cameraType='front'] - Camera type to use
    * @param {boolean} [params.enableAudio=true] - Whether to enable audio
    * @param {boolean} [params.enableVideo=true] - Whether to enable video
@@ -383,6 +384,7 @@ class CustomTwilioVideoView extends Component {
   connect({
     roomName,
     accessToken,
+    region,
     cameraType = "front",
     enableAudio = true,
     enableVideo = true,
@@ -396,6 +398,7 @@ class CustomTwilioVideoView extends Component {
     this.runCommand(nativeEvents.connectToRoom, [
       roomName,
       accessToken,
+      region,
       enableAudio,
       enableVideo,
       enableRemoteAudio,