+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 3.5.0 (In Progress)
+
+### Changes
+
+- Added `region` connect option for both Android and iOS to specify the [Twilio Signaling Region](https://www.twilio.com/docs/video/tutorials/video-regions-and-global-low-latency) (e.g., `us1`, `us2`, `au1`, `br1`, `de1`, `ie1`, `in1`, `jp1`, `sg1`). The `onRoomFetched` callback will now return the local participant's selected region via the `signalingRegion` parameter.
+- Added a region selector dropdown to the Example App main menu for testing different Twilio Signaling Regions.
+
+### Known issues
+
+- Screensharing on iOS only supports in-app sharing. The screen share track will freeze when the app is backgrounded.
+
 ## 3.4.1
 
 ### Changes